# api/openapi.yaml
openapi: 3.0.3
info:
  title: Dockyard Admin API
  description: |
    Administration API of Dockyard, a self-hosted Docker Registry V2.

    Authentication: obtain an access token (15 min) plus a rotating refresh
    token via `POST /api/admin/auth/login`, then send
    `Authorization: Bearer <token>`. The registry protocol itself lives under
    `/v2/*` (Docker Registry HTTP API V2) and authenticates through
    `GET /v2/token` (Docker token auth) — only the token endpoint is described
    here.

    Roles: `admin` (everything), `pusher` (pull + push), `reader` (read-only).
    Endpoints marked *admin only* return `403` for other roles.
  version: "1.0"
  license:
    name: MIT
servers:
  - url: /
...
security:
  - bearerAuth: []
paths:
  /api/admin/auth/login:
    post:
      tags: [auth]
      summary: Log in
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string }
                password: { type: string }
      responses:
        "200":
          description: Access + refresh tokens
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthTokens" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "429": { description: Rate limited (RATE_LIMIT_LOGIN_PER_MIN) }
  /api/admin/auth/refresh:
    post:
      tags: [auth]
      summary: Renew the access token
      description: Refresh tokens are single-use — each call rotates it.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token: { type: string }
      responses:
        "200":
          description: Fresh access + refresh tokens
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthTokens" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/admin/auth/logout:
    post:
      tags: [auth]
      summary: Log out
      description: Revokes the access token (persisted) and kills the session.
      responses:
        "200": { $ref: "#/components/responses/Message" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/admin/auth/password:
    post:
      tags: [auth]
      summary: Change own password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [current_password, new_password]
              properties:
                current_password: { type: string }
                new_password: { type: string, minLength: 8 }
      responses:
        "200": { $ref: "#/components/responses/Message" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/admin/users:
    get:
      tags: [users]
      summary: List users (admin only)
      responses:
        "200":
          description: Users
          content:
            application/json:
              schema:
                type: object
                properties:
                  users:
                    type: array
                    items: { $ref: "#/components/schemas/User" }
                  count: { type: integer }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      tags: [users]
      summary: Create a user (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password, role]
              properties:
                username: { type: string }
                password: { type: string, minLength: 8 }
                role: { $ref: "#/components/schemas/Role" }
                repo_patterns:
                  type: array
                  items: { type: string }
                  description: Repo globs restricting the account ('*' crosses '/').
      responses:
        "201":
          description: Created user
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { description: Username already exists }
  /api/admin/users/{username}:
    parameters:
      - { name: username, in: path, required: true, schema: { type: string } }
    put:
      tags: [users]
      summary: Update role, patterns and/or password (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                role: { $ref: "#/components/schemas/Role" }
                repo_patterns:
                  type: array
                  items: { type: string }
                password: { type: string, minLength: 8 }
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "400": { description: Invalid change (e.g. demoting the last admin) }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [users]
      summary: Delete a user (admin only)
      description: Self-deletion and removing the last admin are refused.
      responses:
        "200": { $ref: "#/components/responses/Message" }
        "400": { description: Refused (own account or last admin) }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/admin/sessions:
    get:
      tags: [sessions]
      summary: List active sessions (admin only)
      responses:
        "200":
          description: Sessions
          content:
            application/json:
              schema:
                type: object
                properties:
                  sessions:
                    type: array
                    items: { $ref: "#/components/schemas/Session" }
                  count: { type: integer }
                  current_id:
                    type: integer
                    description: Session id of the caller.
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/admin/sessions/{id}:
    delete:
      tags: [sessions]
      summary: Revoke a session (admin only)
      description: Kills the refresh token; outstanding access tokens die within 15 minutes.
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200": { $ref: "#/components/responses/Message" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/admin/repositories:
    get:
      tags: [repositories]
      summary: List repositories with tags and last push time
      responses:
        "200":
          description: Repositories
          content:
            application/json:
              schema:
                type: object
                properties:
                  repositories:
                    type: array
                    items:
                      type: object
                      properties:
                        name: { type: string }
                        tags: { type: array, items: { type: string } }
                        total: { type: integer }
                        last_pushed: { type: string, format: date-time }
                  total: { type: integer }
    delete:
      tags: [repositories]
      summary: Delete a repository and all its tags (admin only)
      parameters:
        - { name: name, in: query, required: true, schema: { type: string } }
      responses:
        "200": { $ref: "#/components/responses/Message" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/admin/repositories/tags:
    get:
      tags: [repositories]
      summary: List tags of a repository with digests
      parameters:
        - { name: name, in: query, required: true, schema: { type: string } }
      responses:
        "200":
          description: Tags
          content:
            application/json:
              schema:
                type: object
                properties:
                  name: { type: string }
                  tags:
                    type: array
                    items:
                      type: object
                      properties:
                        tag: { type: string }
                        digest: { type: string }
                        pushed_at: { type: string, format: date-time }
                  total: { type: integer }
  /api/admin/repositories/manifest:
    get:
      tags: [repositories]
      summary: Manifest details (size, layers, platforms)
      parameters:
        - { name: name, in: query, required: true, schema: { type: string } }
        - { name: reference, in: query, required: true, schema: { type: string }, description: Tag or digest }
      responses:
        "200":
          description: Manifest details
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ManifestDetails" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/admin/repositories/diff:
    get:
      tags: [repositories]
      summary: Compare layers/config between two tags or digests of the same image
      parameters:
        - { name: name, in: query, required: true, schema: { type: string } }
        - { name: reference_a, in: query, required: true, schema: { type: string }, description: Tag or digest }
        - { name: reference_b, in: query, required: true, schema: { type: string }, description: Tag or digest }
      responses:
        "200":
          description: Diff
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TagDiff" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/admin/repositories/search:
    get:
      tags: [repositories]
      summary: Search repo:tag pairs by name/tag substring, optionally filtered by signed status
      parameters:
        - { name: q, in: query, schema: { type: string }, description: Case-insensitive substring match on repo name or tag }
        - { name: signed, in: query, schema: { type: string, enum: ["true", "false"] }, description: Only present when cosign public keys are configured }
        - { name: limit, in: query, schema: { type: integer, default: 50, maximum: 200 } }
        - { name: offset, in: query, schema: { type: integer, default: 0 } }
      responses:
        "200":
          description: Matches
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/SearchResult" }
                  total: { type: integer }
                  count: { type: integer }
  /api/admin/repositories/layer:
    get:
      tags: [repositories]
      summary: List the files inside a layer (gzip, zstd or plain tar)
      parameters:
        - { name: name, in: query, required: true, schema: { type: string } }
        - { name: digest, in: query, required: true, schema: { type: string } }
      responses:
        "200":
          description: Layer entries
          content:
            application/json:
              schema:
                type: object
                properties:
                  digest: { type: string }
                  entries:
                    type: array
                    items:
                      type: object
                      properties:
                        path: { type: string }
                        type: { type: string, enum: [file, dir, symlink, hardlink, other] }
                        size: { type: integer }
                        size_human: { type: string }
                        mode: { type: string }
                        link_name: { type: string }
                  count: { type: integer }
  /api/admin/repositories/manifests:
    delete:
      tags: [repositories]
      summary: Delete a manifest (admin only)
      description: Removes the manifest and every tag pointing at its digest.
      parameters:
        - { name: name, in: query, required: true, schema: { type: string } }
        - { name: digest, in: query, required: true, schema: { type: string } }
      responses:
        "200": { $ref: "#/components/responses/Message" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/admin/repositories/export:
    get:
      tags: [repositories]
      summary: Export a repository as an OCI image-layout tarball (admin only)
      description: Interoperates with skopeo/crane. Each tag becomes an index.json entry.
      parameters:
        - { name: name, in: query, required: true, schema: { type: string } }
      responses:
        "200":
          description: Tar stream
          content:
            application/x-tar:
              schema: { type: string, format: binary }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/admin/repositories/import:
    post:
      tags: [repositories]
      summary: Import an OCI image-layout tarball into a repository (admin only)
      parameters:
        - { name: name, in: query, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/x-tar:
            schema: { type: string, format: binary }
      responses:
        "200":
          description: Import result
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
                  tags: { type: integer }
        "400": { description: Not an OCI image layout }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/admin/storage/stats:
    get:
      tags: [storage]
      summary: Storage usage
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                type: object
                properties:
                  total_size_bytes: { type: integer }
                  total_size_human: { type: string }
                  blob_count: { type: integer }
                  repo_count: { type: integer }
                  storage_path: { type: string }
        "501": { description: Not available in proxy mode }
  /api/admin/storage/tree:
    get:
      tags: [storage]
      summary: Raw filesystem tree (local backend only)
      responses:
        "200": { description: File list }
        "501": { description: Not available (proxy mode or S3 backend) }
  /api/admin/gc:
    post:
      tags: [storage]
      summary: Garbage-collect unreferenced blobs (admin only)
      parameters:
        - name: dryRun
          in: query
          schema: { type: boolean }
          description: Report what would be removed without deleting.
      responses:
        "200":
          description: GC result
          content:
            application/json:
              schema:
                type: object
                properties:
                  removed: { type: array, items: { type: string } }
                  count: { type: integer }
                  freed_bytes: { type: integer }
                  freed_human: { type: string }
                  dry_run: { type: boolean }
        "403": { $ref: "#/components/responses/Forbidden" }
        "501": { description: Not available in proxy mode }
  /api/admin/insights:
    get:
      tags: [storage]
      summary: Growth history and largest repositories (admin only)
      responses:
        "200":
          description: Insights
          content:
            application/json:
              schema:
                type: object
                properties:
                  history:
                    type: array
                    items:
                      type: object
                      properties:
                        at: { type: string, format: date-time }
                        total_size: { type: integer }
                        blob_count: { type: integer }
                        repo_count: { type: integer }
                  top_repos:
                    type: array
                    items:
                      type: object
                      properties:
                        name: { type: string }
                        size_bytes: { type: integer }
                        size_human: { type: string }
                        tags: { type: integer }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/admin/retention:
    get:
      tags: [retention]
      summary: List retention policies (admin only)
      responses:
        "200":
          description: Policies
          content:
            application/json:
              schema:
                type: object
                properties:
                  policies:
                    type: array
                    items: { $ref: "#/components/schemas/RetentionPolicy" }
                  count: { type: integer }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      tags: [retention]
      summary: Create a retention policy (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RetentionPolicyInput" }
      responses:
        "201":
          description: Created policy
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RetentionPolicy" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/admin/retention/{id}:
    delete:
      tags: [retention]
      summary: Delete a retention policy (admin only)
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200": { $ref: "#/components/responses/Message" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/admin/retention/run:
    post:
      tags: [retention]
      summary: Evaluate (and optionally apply) the retention plan (admin only)
      parameters:
        - name: dryRun
          in: query
          schema: { type: boolean }
          description: Preview the plan without deleting anything.
      responses:
        "200":
          description: Plan and outcome
          content:
            application/json:
              schema:
                type: object
                properties:
                  plan:
                    type: object
                    properties:
                      delete:
                        type: array
                        items:
                          type: object
                          properties:
                            repo: { type: string }
                            tag: { type: string }
                            digest: { type: string }
                            reason: { type: string }
                      skipped:
                        type: array
                        items:
                          type: object
                          properties:
                            repo: { type: string }
                            tag: { type: string }
                            reason: { type: string }
                  dry_run: { type: boolean }
                  deleted: { type: integer }
        "403": { $ref: "#/components/responses/Forbidden" }
...
  /api/admin/webhooks/{id}:
    delete:
      tags: [webhooks]
      summary: Delete a webhook (admin only)
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200": { $ref: "#/components/responses/Message" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/admin/webhooks/{id}/test:
    post:
      tags: [webhooks]
      summary: Send a synchronous test event (admin only)
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200": { $ref: "#/components/responses/Message" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "502": { description: Endpoint rejected the delivery }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Message:
      description: Success message
      content:
        application/json:
          schema:
            type: object
            properties:
              message: { type: string }
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: Missing, invalid or revoked credentials
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
...
    Role:
      type: string
      enum: [admin, pusher, reader]
    AuthTokens:
      type: object
      properties:
        token: { type: string, description: JWT access token (15 min) }
        refresh_token: { type: string, description: Single-use refresh token (30 days) }
        role: { $ref: "#/components/schemas/Role" }
        expires_in: { type: integer, example: 900 }
...